initial solution for HW2
will have to edit this later; the work done here was under 'duress' (i.e. in
room Ge), and i could not check if numpy slows things down as prof. Silvestri
told me. as it stands the code is _extremely_ slow, but works -- aside from
the artificial dataset, that is; i got tired of waiting.